+++ changelog.md
@@ -1,9 +1,14 @@
+### 06.04.2023, Thu
+
+- 22:00 - Basic Telegram bot features implemented.
+
 ### 05.04.2023, Wed
 
 - 12:50 - Created `users` app with `CustomUser` model.
 - 14:00 - Created `bookmarks` app with `Folder` and `Bookmark` models.
 - 19:00 - Built DRF API endpoints `bookmarks` and `bookmarks/create_from_telegram`.
 - 19:30 - Simple page "meta" content parser added.
+- 22:30 - Telegram bot prototype created.
 
 ### 03.04.2023, Mo